requirements.txt
pandas>=1.5
requests>=2.28
sqlalchemy>=1.4
psycopg2-binary>=2.9
streamlit>=1.24
plotly>=5.15
python-dotenv>=1.0
